+++ CHANGELOG.md
@@ -8,7 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - Viewer: optionally show rating & tags on overlay
 - Viewer: long press on rating quick action for quicker rating
-- Search: missing address filter
+- Search: missing address, portrait, landscape filters
 - Lithuanian translation (thanks Gediminas Murauskas)
 
 ### Changed